REntalService correction

+++ src/main/java/pl/PJATK/RentalService.java
@@ -40,7 +40,7 @@ public class RentalService {
 
     public static double estimatedPrice(String vin, LocalDate rentalDate, LocalDate returnDate){
         CarStorage carstorage = CarStorage.getInstance();
-        double daysOfRent = ChronoUnit.DAYS.between(rentalDate, returnDate);
+        long daysOfRent = ChronoUnit.DAYS.between(rentalDate, returnDate);
         double pricePerDay = 100.0;
         for (Car currentCar : carstorage.getAllCars()){
             if (vin.equals(currentCar.getVin())){
